requirements.txt
fastapi
uvicorn
requests
python-docx
PyMuPDF  # for reading PDFs